More development work on CAFEParam and CAFEState.
Now, the Configuration class has its internal data
represented by a CAFEParam member variable.  In addition,
several member functions were eliminated due to
incompatibilities between the legacy data representation
and how CAFEParam represents its data.  Because of this,
AnomalyCollection and AnomalyCollection_GR had to be
modified because it used an eliminated function.

This is the first step in the gradual integration of the
CAFEParam class into CAFE -- eventually utilizing
CAFEState as the loop controller and downsizing
Configuration and CAFE_CmdLine down to strictly file and
command-line processing.


git-svn-id: svn+ssh://localhost/home/bvr/svnrepos/Programs/CAFE@28 c770f1bc-0775-48ed-bd74-6edc8452e29f