Major development and integration work done for the CAFEState module.
CAFEState is now treated as an advanced logical
association handler.  It now can examine the CAFEParams
from the perspective of the raw CAFEVars or from the
EventVars.  In addition, it is easy to change the
data source that CAFEState uses for multi-source
processing.

CAFEState is now integrated into all (but two) main
programs.  Only AnomalyCollection_GR and TotalRecall_GR
have not been changed, pending continued refactoring of
AnomalyCollection and TotalRecall.

The goal of the integration is to remove the use of
the Configuration and CmdOptions modules from the main
body of these programs and utility functions.  They should
only be used for initial loading of command-line options
and configuration file information.  CAFEState is to
serve as the de facto source of any CAFE-related parameters.
Lots more testing and work is needed for this module, however.
Some Utility modules were also modified to eliminate inappropriate
use of the Configuration and CmdOptions modules.

In addition, BoardConvertor.C was modified to include more
diagnostic information in some error messages.


git-svn-id: svn+ssh://localhost/home/bvr/svnrepos/Programs/CAFE@31 c770f1bc-0775-48ed-bd74-6edc8452e29f